+++ RealmSwift/Realm.swift
@@ -20,6 +20,9 @@ import Foundation
 import Realm
 import Realm.Private
 
+/// The Id of the asynchronous transaction.
+public typealias AsyncTransactionId = RLMAsyncTransactionId
+
 /**
  A `Realm` instance (also referred to as "a Realm") represents a Realm database.
 
@@ -364,6 +367,111 @@ import Realm.Private
      */
     public var isInWriteTransaction: Bool {
         return rlmRealm.inWriteTransaction
+    }
+
+// MARK: Asynchronous Transactions
+
+    /**
+     Asynchronously performs actions contained within the given block inside a write transaction.
+     The write transaction is begun asynchronously as if calling `beginAsyncWrite`,
+     and by default the transaction is commited asynchronously after the block completes.
+     You can also explicitly call `commitWrite` or `cancelWrite` from
+     within the block to synchronously commit or cancel the write transaction.
+     Returning without one of these calls is equivalent to calling `commitWrite`.
+
+     @param block The block containing actions to perform.
+
+     @param completionBlock A block which will be called on the source thread or queue
+                        once the commit has either completed or failed with an error.
+
+     @return An id identifying the asynchronous transaction which can be passed to
+             `cancelAsyncWrite` prior to the block being called to cancel
+             the pending invocation of the block.
+    */
+    @discardableResult
+    public func writeAsync(_ block: @escaping () -> Void, onComplete: ((Swift.Error?) -> Void)? = nil) -> AsyncTransactionId {
+        return beginAsyncWrite {
+            block()
+            commitAsyncWrite(onComplete)
+        }
+    }
+
+    /**
+     Begins an asynchronous write transaction.
+     This function asynchronously begins a write transaction on a background
+     thread, and then invokes the block on the original thread or queue once the
+     transaction has begun. Unlike `beginWrite`, this does not block the
+     calling thread if another thread is current inside a write transaction, and
+     will always return immediately.
+     Multiple calls to this function (or the other functions which perform
+     asynchronous write transactions) will queue the blocks to be called in the
+     same order as they were queued. This includes calls from inside a write
+     transaction block, which unlike with synchronous transactions are allowed.
+
+     @param asyncWriteBlock The block containing actions to perform inside the write transaction.
+            `asyncWriteBlock` should end by calling `commitAsyncWrite` or `commitWrite`.
+            Returning without one of these calls is equivalent to calling `cancelAsyncWrite`.
+
+     @return An id identifying the asynchronous transaction which can be passed to
+             `cancelAsyncWrite` prior to the block being called to cancel
+             the pending invocation of the block.
+     */
+    @discardableResult
+    public func beginAsyncWrite(_ asyncWriteBlock: @escaping () -> Void) -> AsyncTransactionId {
+        return rlmRealm.beginAsyncWriteTransaction {
+            asyncWriteBlock()
+        }
+    }
+
+    /**
+     Asynchronously commits a write transaction.
+     The call returns immediately allowing the caller to proceed while the I/O is
+     performed on a dedicated background thread. This can be used regardless of if
+     the write transaction was begun with `beginWrite` or `beginAsyncWrite`.
+
+     @param onComplete A block which will be called on the source thread or queue once the commit
+                     has either completed or failed with an error.
+
+     @param isGroupingAllowed If `true`, multiple sequential calls to `commitAsyncWrite` may be
+                          batched together and persisted to stable storage in one group. This
+                          improves write performance, particularly when the individual transactions
+                          being batched are small. In the event of a crash or power failure,
+                          either all of the grouped transactions will be lost or none will, rather
+                          than the usual guarantee that data has been persisted as
+                          soon as a call to commit has returned.
+
+     @return An id identifying the asynchronous transaction commit can be passed to
+             `cancelAsyncWrite` prior to the completion block being called to cancel
+             the pending invocation of the block. Note that this does *not* cancel the commit itself.
+    */
+    @discardableResult
+    public func commitAsyncWrite(_ onComplete: ((Swift.Error?) -> Void)? = nil, isGroupingAllowed: Bool = false) -> AsyncTransactionId {
+        return rlmRealm.commitAsyncWriteTransaction(onComplete, isGroupingAllowed: isGroupingAllowed)
+    }
+
+    /**
+     Cancels a queued block for an asynchronous transaction.
+     This can cancel a block passed to either an asynchronous begin or an asynchronous commit.
+     Canceling a begin cancels that transaction entirely, while canceling a commit merely cancels
+     the invocation of the completion callback, and the commit will still happen.
+     Transactions can only be canceled before the block is invoked, and calling `cancelAsyncWrite`
+     from within the block is a no-op.
+
+     @param AsyncTransactionId A transaction id from either `beginAsyncWrite` or `commitAsyncWrite`.
+    */
+    public func cancelAsyncWrite(_  asyncTransactionId: AsyncTransactionId) throws {
+        rlmRealm.cancelAsyncTransaction(asyncTransactionId)
+    }
+
+    /**
+     Indicates if the Realm is currently performing async write operations.
+     This becomes `true` following a call to `beginAsyncWrite`, `commitAsyncWrite`,
+     or `writeAsync`, and remains so until all scheduled async write work has completed.
+
+     @warning If this is `true`, closing or invalidating the Realm will block until scheduled work has completed.
+     */
+    public var isPerformingAsynchronousWriteOperations: Bool {
+        return rlmRealm.isPerformingAsynchronousWriteOperations
     }
 
     // MARK: Adding and Creating objects