+++ CHANGELOG.md
@@ -1,6 +1,24 @@
 x.y.z Release notes (yyyy-MM-dd)
 =============================================================
 ### Enhancements
+* Add Swift API for asynchronous transactions
+```
+    try? realm.writeAsync {
+        realm.create(SwiftStringObject.self, value: ["string"])
+    } onComplete: { error in
+        // optional handling on write complete
+    }
+
+    try? realm.beginAsyncWrite {
+        realm.create(SwiftStringObject.self, value: ["string"])
+        realm.commitAsyncWrite()
+    }
+
+    let asyncTransactionId = try? realm.beginAsyncWrite {
+        // ...
+    }
+    try! realm.cancelAsyncWrite(asyncTransactionId)
+```
 * Add Obj-C API for asynchronous transactions
 ```
    [realm asyncTransactionWithBlock:^{